+++ internal/tools/common.go
@@ -1,0 +1,74 @@
+package tools
+
+import (
+	"fmt"
+	"regexp"
+	"strings"
+
+	"github.com/xuri/excelize/v2"
+)
+
+// parseRange parses Excel's range string (e.g. A1:C10)
+func ParseRange(rangeStr string) (int, int, int, int, error) {
+	re := regexp.MustCompile(`(\$?[A-Z]+\$?\d+):(\$?[A-Z]+\$?\d+)`)
+	matches := re.FindStringSubmatch(rangeStr)
+	if matches == nil {
+		return 0, 0, 0, 0, fmt.Errorf("invalid range format: %s", rangeStr)
+	}
+	startCol, startRow, err := excelize.CellNameToCoordinates(matches[1])
+	if err != nil {
+		return 0, 0, 0, 0, err
+	}
+	endCol, endRow, err := excelize.CellNameToCoordinates(matches[2])
+	if err != nil {
+		return 0, 0, 0, 0, err
+	}
+	return startCol, startRow, endCol, endRow, nil
+}
+
+// CreateHTMLTable creates a table data in HTML format
+func CreateHTMLTable(workbook *excelize.File, sheetName string, startCol int, startRow int, endCol int, endRow int) (*string, error) {
+	var table string
+	table += "<table>\n"
+
+	// ヘッダー行（範囲情報）
+	startCell, err := excelize.CoordinatesToCellName(startCol, startRow)
+	if err != nil {
+		return nil, err
+	}
+	endCell, err := excelize.CoordinatesToCellName(endCol, endRow)
+	if err != nil {
+		return nil, err
+	}
+	responseRange := fmt.Sprintf("%s:%s", startCell, endCell)
+	fullRange, err := workbook.GetSheetDimension(sheetName)
+	if err != nil {
+		return nil, err
+	}
+	table += fmt.Sprintf("<tr><th>[%s] Current data range: %s, Full data range: %s</th>",
+		sheetName, responseRange, fullRange)
+
+	// 列アドレスの出力
+	for col := startCol; col <= endCol; col++ {
+		name, _ := excelize.ColumnNumberToName(col)
+		table += fmt.Sprintf("<th>%s</th>", name)
+	}
+	table += "</tr>\n"
+
+	// データの出力
+	for row := startRow; row <= endRow; row++ {
+		table += "<tr>"
+		// 行アドレスを出力
+		table += fmt.Sprintf("<td>%d</td>", row)
+
+		for col := startCol; col <= endCol; col++ {
+			axis, _ := excelize.CoordinatesToCellName(col, row)
+			value, _ := workbook.GetCellValue(sheetName, axis)
+			table += fmt.Sprintf("<td>%s</td>", strings.ReplaceAll(value, "\n", "<br>"))
+		}
+		table += "</tr>\n"
+	}
+
+	table += "</table>"
+	return &table, nil
+}
